added: Welcome Screen complete
